rework transformers exports
