--- src/attributes/primitive/constants.ts
@@ -1,5 +1,0 @@
-export const $transformerId = Symbol('$transformerId')
-export type $transformerId = typeof $transformerId
-
-export const $typeModifier = Symbol('$typeModifier')
-export type $typeModifier = typeof $typeModifier
