+++ src/schema/actions/parse/primitive.ts
@@ -14,12 +14,11 @@ import type {
   ResolveStringAttribute,
   ResolvedNullAttribute,
   ResolvedPrimitiveAttribute,
-  StringAttribute,
-  Transformer,
-  TypeModifier
+  StringAttribute
 } from '~/attributes/index.js'
 import { DynamoDBToolboxError } from '~/errors/index.js'
 import type { Schema } from '~/schema/index.js'
+import type { Transformer, TypeModifier } from '~/transformers/index.js'
 import type { If } from '~/types/index.js'
 import { cloneDeep } from '~/utils/cloneDeep.js'
 import { isValidPrimitive } from '~/utils/validation/isValidPrimitive.js'
