+++ src/transformers/transformer.ts
@@ -1,0 +1,34 @@
+import type { Constant, Fn, Identity } from 'hotscript'
+
+export interface Transformer<
+  FORMATTED_CONSTRAINT = any,
+  FORMATTED extends FORMATTED_CONSTRAINT = FORMATTED_CONSTRAINT,
+  TRANSFORMED = any
+> {
+  parse: (formatted: FORMATTED) => TRANSFORMED
+  format: (transformed: TRANSFORMED) => FORMATTED_CONSTRAINT
+}
+
+export interface TypedTransformer<
+  FORMATTED_CONSTRAINT = any,
+  FORMATTED extends FORMATTED_CONSTRAINT = FORMATTED_CONSTRAINT,
+  TRANSFORMED = any,
+  TYPE_MODIFIER extends Fn = Identity
+> extends Transformer<FORMATTED_CONSTRAINT, FORMATTED, TRANSFORMED> {
+  _typeModifier: TYPE_MODIFIER
+}
+
+export type TypeModifier<TRANSFORMER extends Transformer> = TRANSFORMER extends TypedTransformer
+  ? TRANSFORMER['_typeModifier']
+  : Constant<ReturnType<TRANSFORMER['parse']>>
+
+export interface JSONizableTransformer<
+  FORMATTED_CONSTRAINT = any,
+  FORMATTED extends FORMATTED_CONSTRAINT = FORMATTED_CONSTRAINT,
+  TRANSFORMED = any,
+  TYPE_MODIFIER extends Fn = Fn,
+  JSONIZED extends { transformerId: string } & object = { transformerId: string } & object
+> extends TypedTransformer<FORMATTED_CONSTRAINT, FORMATTED, TRANSFORMED, TYPE_MODIFIER> {
+  transformerId: JSONIZED['transformerId']
+  jsonize: () => JSONIZED
+}
